+++ ReleaseNotes/release_notes.md
@@ -1,3 +1,4 @@
+# v3.6.9.14
 
 ## New features
 
@@ -6,6 +7,11 @@
   1. Add Dixon elliptic functions to `ostap.math.math_ve`
   1. Add lemniscate elliptic sine and cosine functions
   1. Add lemniscate elliptic hyperbolic sine and cosine functions
+  1. Add complex eta and zeta functions
+  1. reshuffle `Ostap/MoreMath.h` header fiel ans split it into smaller pieces 
+  1. Add Hadamard's (pseudo) gamma function
+  1. Add Luschny's  (L)-factorial function
+
 
 # v3.6.9.12